Update docs content to match recent updates
