+++ docs/content/guides/initial-setup.md
@@ -15,15 +15,18 @@ tags:
 A project created on Fireadmin can contain multiples environments which represent multiple versions of your project. The environment names usually correspond with the state of the application within that environment (i.e. prod for a production environment and stage for a staging environment). In this case each environment is associated with a separate Firebase project.
 
 ## Before Starting
+
 1. Make sure you have multiple projects created on Firebase to represent your multiple environment. We will be running an action to move data between them.
 1. Get a service a service account for each Firebase project you will be adding to Fireadmin. To get a service account navigate to the Service Accounts tab in the Firebase Console “project settings” page. Then click Generate New Private Key at the bottom of the Service Accounts tab. Your service account will be downloaded as a JSON file — rename the downloaded files in a way that makes it easy to tell a difference (i.e. service-account-prod.json).
 
 ## Creating a Project
+
 1. Create a project by clicking the add tile — we will use the name firething (Note: remember that projects on Fireadmin usually contain multiple Firebase projects)
 
 ![image](https://cdn-images-1.medium.com/max/1000/1*36v3pDNR07JQkBDffGMAcg.gif)
 
 ## Adding Environments
+
 1. Go to Environments page by clicking Go To Environments or by clicking Environments in the project sidebar
 1. Begin adding a new environment by clicking Add Environment
 1. You can name the first environment whatever you want, but I usually go with Production or Prod
@@ -35,8 +38,8 @@ A project created on Fireadmin can contain multiples environments which represen
 
 ![image](https://cdn-images-1.medium.com/max/1000/1*abIK5N_ZDwtaC-rlXP3ZUw.gif)
 
-
 ## Running An Action
+
 Now that we have environments setup, they are ready to be used in Actions (common logic such as copying/moving data). Lets start by searching through the existing action templates.
 
 1. Go to the Project Actions page by clicking Actions in the Sidebar
@@ -49,4 +52,4 @@ Now that we have environments setup, they are ready to be used in Actions (commo
 
 ![Running An Action](https://cdn-images-1.medium.com/max/1000/1*N62nuM6LJP-7xTV2oP0F_Q.gif)
 
-For information on how to creat your own action template, checkout the [custom action template guide](/guides/custom-action-templates)
+For information on how to create your own action template, checkout the [custom action template guide](/guides/custom-action-templates)
