+++ docs/content/source/file-structure.md
@@ -13,8 +13,8 @@ Placement of files within a project repository is important, especially as the p
 
 ##### Location
 
-* Global: `src/components`
-* Contained/In one route only: `src/routes/:routeName/components`
+- Global: `src/components`
+- Contained/In one route only: `src/routes/:routeName/components`
 
 ##### Creation
 
@@ -26,25 +26,21 @@ yo react-firebase:component
 
 ##### Handles
 
-* All view layout/styling
+- All view layout/styling
 
 ##### Patterns
 
-* Should be stateless \(just a function instead of a Class\)
+- Should be stateless \(just a function instead of a Class\)
 
 ##### Example
 
 `src/components/SomeComponent.js`:
 
 ```js
-import React, { PropTypes } from 'react';
+import React, { PropTypes } from "react";
 
 function SomeComponent({ someProp }) {
-  return (
-    <div>
-      Some Component
-    </div>
-  );
+  return <div>Some Component</div>;
 }
 
 SomeComponent.propTypes = {
@@ -56,40 +52,33 @@ export default SomeComponent;
 
 #### Approved file names by component
 
-These are the only authorized files, anything that doesn't have a clear home should go in \*enhancer.js.  **_(Need to update the generator)_**
+These are the only authorized files, anything that doesn't have a clear home should go in \*enhancer.js. **_(Need to update the generator)_**
 
-| Name | Purpose |
-| --- | --- |
-|	`Component.connect.js`	|	Data structuring. Connecting to firebase/firestore/redux and used for decorators like `withProps`	|
-|	`Component.constants.js`	|	Contants specific to this component	|
-|	`Component.enhancer.js`	|	Adds data and handling to the Component, importing the following files to do so: (ALSO THE CATCHALL)	|
-|	`Component.handlers.js`	|	Action functions	|
-|	`Component.js`	|	Functional React Component	|
-|	`Component.lifecycle.js`	|	All lifecycle methods	|
-|	`Component.readme.md`	|		|
-|	`Component.selectors.js`	|	Reselect (re-reselect?) selectors - pure javascript - no react	|
-|	`Component.spec.js`	|	Unit tests per component	|
-|	`Component.stateHandlers.js`	|	Internal component state	|
-|	`Component.stories.js`	|		|
-|	`Component.styles.js`	|	JS Styles / MUI V1 stylesheets	|
-|	`Component.utils.js`	|	Utils file	|
-|	`index.js`	|	Exports the enhanced Component	|
+| Name                    | Purpose                                                                                              |
+| ----------------------- | ---------------------------------------------------------------------------------------------------- |
+| `Component.enhancer.js` | Adds data and handling to the Component, importing the following files to do so: (ALSO THE CATCHALL) |
+| `Component.js`          | Functional React Component                                                                           |
+| `Component.spec.js`     | Unit tests per component                                                                             |
+| `Component.styles.js`   | JS Styles / MUI V1 stylesheets                                                                       |
+| `index.js`              | Exports the enhanced Component                                                                       |
 
 ### Routes
 
 Async vs Sync definitions
 
 ##### Location
-* Route Folder: `src/routes/SomeRoute`
-* Components: `src/routes/SomeRoute/components/SomeComponent/SomeComponent.js`
+
+- Route Folder: `src/routes/SomeRoute`
+- Components: `src/routes/SomeRoute/components/SomeComponent/SomeComponent.js`
 
 ##### Creation
-Create a new container name "SomeRoute" by running \`redux g container SomeRoute\`
+
+Create a new container name "SomeRoute" by running `yo react-firebase:route SomeRoute`
 
 ##### What they Handle
 
-* Pages/Routes/Locations within app
-* They can contain components, and modules
+- Pages/Routes/Locations within app
+- They can contain components, and modules
 
 #### Children
 
@@ -100,7 +89,6 @@ Create a new container name "SomeRoute" by running \`redux g container SomeRoute
 ##### Routes
 
 `index.js`
-
 
 ## Assets
 
@@ -117,9 +105,9 @@ Import in javascript, then use import as ref
 To import `asset.png`:
 
 ```js
-import Asset from 'static/asset.png';
+import Asset from "static/asset.png";
 // Then later
-<img src={Asset} />
+<img src={Asset} />;
 ```
 
 ##### Why
@@ -130,18 +118,8 @@ This allows Webpack to handle chunking of asset files, which means in the case o
 
 ##### Location
 
-Global/Shared: `src/styles`
+Global/Shared: `src/theme`
 
 ##### Patterns
 
-* Wrapper class name is `container`
-* import base to get access to shared variables in css modules:
-
-```css
-@import 'base';
-.container {
-  @extend .flex-row;
-}
-```
-
-Some components use the new styles.js pattern - more to come on this once the full switch has happened.
+Styles for components are within `.styles` file within the component folder
