+++ docs/package.json
@@ -28,11 +28,9 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
-    "react-helmet": "^5.2.0",
-    "styled-components": "^4.1.3"
+    "react-helmet": "^5.2.0"
   },
   "devDependencies": {
-    "babel-plugin-styled-components": "^1.10.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.5.0",
@@ -52,7 +50,6 @@
     "gatsby-plugin-react-helmet": "3.0.4",
     "gatsby-plugin-root-import": "^2.0.5",
     "gatsby-plugin-sharp": "2.0.15",
-    "gatsby-plugin-styled-components": "^3.0.5",
     "gatsby-remark-autolink-headers": "^2.0.12",
     "gatsby-remark-copy-linked-files": "2.0.7",
     "gatsby-remark-images": "3.0.1",
