--- docs/src/components/SidebarItem/SidebarItem.enhancer.js
@@ -1,33 +1,0 @@
-import { compose } from 'redux'
-import { withStateHandlers, withProps } from 'recompose'
-import { get, filter } from 'lodash'
-import { Link } from 'gatsby'
-import { slugIsInCurrentPath } from '../../utils'
-
-function getChildChapters(pages) {
-  return filter(pages, (page) => {
-    const slug = get(page, 'node.frontmatter.slug') || ''
-    return slug.split('/').length > 1
-  })
-}
-
-export default compose(
-  withStateHandlers(() => ({ open: false }), {
-    toggleOpen: ({ open }) => () => {
-      return {
-        open: !open
-      }
-    }
-  }),
-  withProps(({ childPages, frontmatter, toggleOpen, location }) => {
-    const childChapters = getChildChapters(childPages)
-    return {
-      childChapters,
-      parentMatchesPath: slugIsInCurrentPath(frontmatter.slug),
-      parentProps:
-        childChapters && childChapters.length
-          ? { onClick: toggleOpen }
-          : { component: Link, to: frontmatter.slug }
-    }
-  })
-)
