+++ docs/src/components/SidebarItem/index.js
@@ -1,4 +1,3 @@
 import SidebarItem from './SidebarItem'
-import enhance from './SidebarItem.enhancer'
 
-export default enhance(SidebarItem)
+export default SidebarItem
